requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.74.0
pyyaml==6.0.1
watchdog==3.0.0
python-dotenv==1.0.0
pytest==7.4.2
asyncio==3.4.3
cryptography==41.0.4
portalocker==2.8.2